nit: rename to mockUseDialogContext and remove unneeded .mock. globs
